+++ js/util.js
@@ -1,0 +1,11 @@
+//функция по получению случайпного числа из диапазона
+const getRandomInteger = (a,b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+//разделяй и властвуй
+const getRandomArrayElement = (elements) => elements [getRandomInteger(0, elements.length - 1)];
+
+export {getRandomArrayElement, getRandomInteger};
